Remove classes that are not use for the moment - move Breakpoints class to debugging session package
